Update PWM phasing logic

+++ Software/Core/Inc/pwm.h
@@ -14,20 +14,20 @@ void pwm_init();
  * @param dutyValue the duty cycle value, 0-2^16
  * @param enableOutput true to enable the output, false to disable the half bridge
  */
-void setPhaseADuty(uint16_t dutyValue, bool enableOutput);
+void setPhaseADuty(uint32_t dutyValue, bool enableOutput);
 
 /**
  * Sets the duty cycle of the PWM signal for Phase A
  * @param dutyValue the duty cycle value, 0-2^16
  * @param enableOutput true to enable the output, false to disable the half bridge
  */
-void setPhaseBDuty(uint16_t dutyValue, bool enableOutput);
+void setPhaseBDuty(uint32_t dutyValue, bool enableOutput);
 
 /**
  * Sets the duty cycle of the PWM signal for Phase A
  * @param dutyValue the duty cycle value, 0-2^16
  * @param enableOutput true to enable the output, false to disable the half bridge
  */
-void setPhaseCDuty(uint16_t dutyValue, bool enableOutput);
+void setPhaseCDuty(uint32_t dutyValue, bool enableOutput);
 
 #endif
